requirements.txt
fastapi
uvicorn
python-dotenv
openai
google-auth
sqlalchemy
psycopg2-binary
stripe